Feature/pgi2 (#1646)
* accommodations for nvphc

* bugfixes

* test_ufo_opr_crtm_amsua passes

* turn FPE back on

* trackcheck tests pass

* clean up

* remove tab

* coding norms

* review comments

+++ src/ufo/crtm/ObsRadianceCRTM.interface.h
@@ -30,7 +30,7 @@ extern "C" {
                                   oops::Variables &);
   void ufo_radiancecrtm_delete_f90(F90hop &);
   void ufo_radiancecrtm_simobs_f90(const F90hop &, const F90goms &, const ioda::ObsSpace &,
-                               const int &, const int &, double &, const F90goms &);
+                               const size_t &, const size_t &, double &, const F90goms &);
 // -----------------------------------------------------------------------------
 
 }  // extern C
